requirements.txt
aiofiles==22.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
coloredlogs==15.0.1
dataclasses-json==0.6.7
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
exceptiongroup==1.3.0
frozenlist==1.7.0
gql==3.5.1
graphql-core==3.2.3
graphql_query==1.4.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.5.0
httpx==0.28.1
httpx-sse==0.4.1
humanfriendly==10.0
hvac==2.3.0
idna==3.10
instructor==1.10.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.71
langchain-ollama==0.3.6
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langsmith==0.4.8
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==5.2.0
mypy==1.15.0
mypy_extensions==1.1.0
numba==0.61.2
numpy==2.1.3
ollama==0.5.1
openai==1.97.1
optype==0.9.3
orjson==3.11.0
packaging==25.0
pandas==2.2.3
pandas-stubs==2.3.0.250703
pluggy==1.0.0
prompt-toolkit==3.0.28
propcache==0.3.2
pydantic==2.10.6
pydantic-settings==2.10.1
pydantic_core==2.27.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
rasa-sdk==3.6.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
ruff==0.11.2
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-routing==0.7.2
scikit-learn==1.6.1
scipy==1.15.1
scipy-stubs==1.15.3.0
shap==0.46.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
typer==0.16.0
types-pytz==2025.2.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.7.0
urllib3==2.5.0
uvloop==0.17.0
wcwidth==0.2.6
websockets==10.4
yarl==1.20.1
zstandard==0.23.0
